requirements.txt
boto3==1.24.40
Flask==2.1.3
python-dotenv==0.20.0
